🙏Buy me a tea☕
We support motherless and fatherless children 🚼 

We support local initiatives and provide education and vocational training for orphans, helping them become self-sufficient in theg future. We believe that even small actions can make a big difference in the lives of orphans.

We are an organization that provides support and services to orphans. Some examples include UNICEF, Save the Children, and SOS Children's Villages.

Please, Volunteer your time at local orphanages or foster care programs. Many organizations rely on volunteers to help us with activities, tutoring, and mentoring.

Consider becoming a foster parent or adopting a child. There are many children in need of a loving home and family.

We do our best to keep smiles on their faces. We started 3 years ago and it's our time to come to show the world full of passion folks. 

We believe in contributing can support the majority.

Your little help can change the world 🌏 

Signed-off-by: Augustin Ayama  <122065729+MarketingIQ@users.noreply.github.com>